Added "mock factory" factory function
